+++ src/main/java/com/customerService/domain/UserInfoExpand.java
@@ -5,8 +5,8 @@ import lombok.Data;
 
 @Data
 public class UserInfoExpand {
-    private Integer id;                 // 主键
-    private Integer infoId;             // 客服信息主表id
+    private Long id;                    // 主键
+    private Long infoId;             // 客服信息主表id
     private String externalNickname;    // 对外显示昵称
     private String email;               // 邮件
     private String phone;               // 电话
